move image search to its own component
