+++ web-ui/src/components/image-search/image-search.tsx
